# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Plexara Health Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List all patient profiles for the authenticated user
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
        "401":
          description: Unauthorized
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a new patient profile
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientBody"
      responses:
        "201":
          description: Patient created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
  /patients/{patientId}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get a patient profile
      security:
        - clerkAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Not found
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update a patient profile
      security:
        - clerkAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePatientBody"
      responses:
        "200":
          description: Patient updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "400":
          description: Invalid input
        "404":
          description: Not found

  # Records endpoints
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    HealthListItem:
      type: object
      description: Generic free-form item used for allergies, medications, conditions.
      additionalProperties:
        type: ["string", "null"]
    Patient:
      type: object
      properties:
        id:
          type: integer
        accountId:
          type: string
        displayName:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        sex:
          type: ["string", "null"]
          enum: [male, female, other, null]
        ethnicity:
          type: ["string", "null"]
        isPrimary:
          type: boolean
        platformConsentAcceptedAt:
          type: ["string", "null"]
          format: date-time
        platformConsentVersion:
          type: ["string", "null"]
        heightCm:
          type: ["integer", "null"]
        weightKg:
          type: ["string", "null"]
        physicianName:
          type: ["string", "null"]
        physicianContact:
          type: ["string", "null"]
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactPhone:
          type: ["string", "null"]
        emergencyContactRelationship:
          type: ["string", "null"]
        allergies:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/HealthListItem"
        medications:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/HealthListItem"
        conditions:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/HealthListItem"
        priorSurgeries:
          type: ["string", "null"]
        priorHospitalizations:
          type: ["string", "null"]
        familyHistory:
          type: ["string", "null"]
        additionalHistory:
          type: ["string", "null"]
        smokingStatus:
          type: ["string", "null"]
        alcoholStatus:
          type: ["string", "null"]
        onboardingTourCompletedAt:
          type: ["string", "null"]
          format: date-time
        relation:
          type: ["string", "null"]
          enum: [owner, collaborator, null]
          description: How the calling account relates to this patient.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - accountId
        - displayName
        - isPrimary
        - createdAt
        - updatedAt
    CreatePatientBody:
      type: object
      properties:
        displayName:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        sex:
          type: ["string", "null"]
          enum: [male, female, other, null]
        ethnicity:
          type: ["string", "null"]
      required:
        - displayName
    UpdatePatientBody:
      type: object
      properties:
        displayName:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        sex:
          type: ["string", "null"]
          enum: [male, female, other, null]
        ethnicity:
          type: ["string", "null"]
        heightCm:
          type: ["integer", "null"]
        weightKg:
          type: ["string", "null"]
        physicianName:
          type: ["string", "null"]
        physicianContact:
          type: ["string", "null"]
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactPhone:
          type: ["string", "null"]
        emergencyContactRelationship:
          type: ["string", "null"]
        allergies:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/HealthListItem"
        medications:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/HealthListItem"
        conditions:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/HealthListItem"
        priorSurgeries:
          type: ["string", "null"]
        priorHospitalizations:
          type: ["string", "null"]
        familyHistory:
          type: ["string", "null"]
        additionalHistory:
          type: ["string", "null"]
        smokingStatus:
          type: ["string", "null"]
        alcoholStatus:
          type: ["string", "null"]
        onboardingTourCompletedAt:
          type: ["string", "null"]
          format: date-time
...